pyproject.toml
[build-system]
requires = ["setuptools>=61,<70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trustifai"
version = "0.1.2"
description = "Trustifai: A Comprehensive Framework for AI Trustworthiness"
readme = "README.md"
authors = [
    { name = "Aaryan Verma" }
]
requires-python = ">=3.10"
dependencies = [
    "ipython==8.12.3",
    "litellm==1.81.1",
    "nest_asyncio==1.6.0",
    "nltk==3.8.1",
    "numpy==2.4.1",
    "pydantic==2.12.5",
    "pytest==8.3.4",
    "python-dotenv==1.2.1",
    "pyvis==0.3.2",
    "PyYAML==6.0.3",
    "tenacity==8.2.3"
]
license = { text = "MIT" }

[project.optional-dependencies]
tracing = ["mlflow==3.8.1"]

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",]

[project.urls]
Homepage = "https://github.com/Aaryanverma/trustifai"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]